+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 UNRELEASED
 ----------
 
+* [ [#1011](https://github.com/digitalfabrik/integreat-cms/issues/1011) ] Use ISO format for datetime objects/fields
+* [ [#1599](https://github.com/digitalfabrik/integreat-cms/issues/1599) ] Exclude automatic saves and pending reviews from XLIFF export
+
 
 2022.9.1
 --------